Another iteration on import images from Ballotpedia.
