+++ import_export_ballotpedia/controllers.py
@@ -2,16 +2,24 @@
 # Brought to you by We Vote. Be good.
 
 # -*- coding: UTF-8 -*-
+from bs4 import BeautifulSoup
+import json
+
+from django.contrib import messages
+
 from .models import BallotpediaApiCounterManager
 from ballot.models import BallotItemListManager, BallotItemManager, BallotReturned, BallotReturnedManager, \
     VoterBallotSavedManager
-from candidate.models import CandidateManager, CandidateListManager, fetch_candidate_count_for_office
+from candidate.controllers import save_image_to_candidate_table
+from candidate.models import CandidateManager, CandidateListManager, fetch_candidate_count_for_office, \
+    PROFILE_IMAGE_TYPE_BALLOTPEDIA, PROFILE_IMAGE_TYPE_UNKNOWN
 from config.base import get_environment_variable
 from electoral_district.models import ElectoralDistrict, ElectoralDistrictManager
 from election.models import BallotpediaElection, ElectionManager, Election
 from exception.models import handle_exception
 from geopy.geocoders import get_geocoder_for_service
-import json
+from image.controllers import IMAGE_SOURCE_BALLOTPEDIA, organize_object_photo_fields_based_on_image_type_currently_active
+from organization.controllers import save_image_to_organization_table
 from measure.models import ContestMeasureListManager, ContestMeasureManager
 from office.models import ContestOfficeListManager, ContestOfficeManager
 from polling_location.models import PollingLocationManager
@@ -19,6 +27,8 @@ import requests
 from voter.models import fetch_voter_id_from_voter_device_link, VoterAddressManager
 import wevote_functions.admin
 from wevote_functions.functions import convert_to_int, extract_state_code_from_address_string, positive_value_exists
+from wevote_settings.models import RemoteRequestHistory, RemoteRequestHistoryManager, \
+    RETRIEVE_POSSIBLE_BALLOTPEDIA_PHOTOS
 
 BALLOTPEDIA_API_KEY = get_environment_variable("BALLOTPEDIA_API_KEY")
 BALLOTPEDIA_API_CANDIDATES_URL = get_environment_variable("BALLOTPEDIA_API_CANDIDATES_URL")
@@ -39,7 +49,9 @@ BALLOTPEDIA_API_SAMPLE_BALLOT_RESULTS_TYPE = "sample_ballot_results"
 GEOCODE_TIMEOUT = 10
 GOOGLE_MAPS_API_KEY = get_environment_variable("GOOGLE_MAPS_API_KEY")
 
+IMG_CLASS_NAME_WE_ARE_SEEKING = "widget-img"
 IMPORT_BALLOT_ITEM = 'IMPORT_BALLOT_ITEM'
+SILHOUETTE_PLACEHOLDER_IMAGE_NAME = "Silhouette_Placeholder_Image"
 
 HEADERS_FOR_BALLOTPEDIA_API_CALL = {
     'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
@@ -146,6 +158,17 @@ PRESIDENTIAL_CANDIDATES_JSON_LIST = [
 ]
 
 logger = wevote_functions.admin.get_logger(__name__)
+
+
+# Retrieves the parsed HTML content from the given URL.
+def get_parsed_html(url):
+    try:
+        page = requests.get(url)
+        return BeautifulSoup(page.content, "html.parser")
+
+    except requests.exceptions.RequestException:
+        print('Unable to connect to {}'.format(url))
+        return BeautifulSoup('', "lxml")
 
 
 def extract_value_from_array(structured_json, index_key, default_value):
@@ -481,6 +504,177 @@ def attach_ballotpedia_election_by_district_from_api(election, google_civic_elec
     }
     return results
 
+
+def get_ballotpedia_photo_url_from_ballotpedia_candidate_url_page(ballotpedia_candidate_url):
+    is_silhouette = False
+    photo_url = ""
+    photo_url_found = False
+    status = ""
+    success = True
+    soup = get_parsed_html(ballotpedia_candidate_url)
+    for img in soup.find_all(class_=IMG_CLASS_NAME_WE_ARE_SEEKING):
+        if photo_url_found:
+            continue
+        photo_url = img.get('src')  # Use get() method to safely retrieve attributes
+        if photo_url:
+            try:
+                print(img['alt'], photo_url)
+                is_silhouette = SILHOUETTE_PLACEHOLDER_IMAGE_NAME in photo_url
+                if is_silhouette:
+                    status += "SILHOUETTE_PLACEHOLDER_FOUND "
+                else:
+                    photo_url_found = True
+            except Exception as e:
+                status += "ERROR_TRYING_TO_GET_BALLOTPEDIA_PHOTO_URL: " + str(e) + " "
+                success = False
+                status += ("Image URL not found for:", img['alt'])
+
+    results = {
+        'is_silhouette':    is_silhouette,
+        'photo_url':        photo_url,
+        'photo_url_found':  photo_url_found,
+        'status':           status,
+        'success':          success,
+    }
+    return results
+
+
+def get_photo_url_from_ballotpedia(
+        incoming_object=None,
+        request={},
+        remote_request_history_manager=None,
+        save_to_database=False,
+        add_messages=False):
+    status = ""
+    success = True
+    ballotpedia_photo_saved = False
+    is_candidate = False
+    is_politician = False
+    if remote_request_history_manager is None:
+        remote_request_history_manager = RemoteRequestHistoryManager()
+
+    if hasattr(incoming_object, 'ballotpedia_candidate_url'):
+        ballotpedia_page_url = incoming_object.ballotpedia_candidate_url
+        google_civic_election_id = incoming_object.google_civic_election_id
+        is_candidate = True
+    elif hasattr(incoming_object, 'ballotpedia_politician_url'):
+        ballotpedia_page_url = incoming_object.ballotpedia_politician_url
+        google_civic_election_id = ''
+        is_politician = True
+    else:
+        ballotpedia_page_url = incoming_object.organization_ballotpedia
+        google_civic_election_id = ''
+        is_candidate = False
+
+    if not positive_value_exists(ballotpedia_page_url):
+        status += "MISSING_BALLOTPEDIA_PAGE_URL "
+        results = {
+            'success': success,
+            'status': status,
+        }
+        return results
+
+    print(ballotpedia_page_url)
+    results = get_ballotpedia_photo_url_from_ballotpedia_candidate_url_page(ballotpedia_page_url)
+    if results.get('success'):
+        photo_url = results.get('photo_url')
+        if save_to_database:
+            # To explore, when photo_url is found, but not valid... (low priority)
+            # ballotpedia_photo_url_is_broken = results.get('http_response_code') == 404
+            if results['photo_url_found']:
+                if is_candidate or is_politician:
+                    incoming_object.ballotpedia_photo_url = photo_url
+                    incoming_object.ballotpedia_photo_url_is_broken = False
+                    if incoming_object.profile_image_type_currently_active == PROFILE_IMAGE_TYPE_BALLOTPEDIA:
+                        incoming_object.profile_image_type_currently_active = PROFILE_IMAGE_TYPE_UNKNOWN
+                        incoming_object.we_vote_hosted_profile_image_url_large = None
+                        incoming_object.we_vote_hosted_profile_image_url_medium = None
+                        incoming_object.we_vote_hosted_profile_image_url_tiny = None
+                        results = organize_object_photo_fields_based_on_image_type_currently_active(
+                            object_with_photo_fields=incoming_object)
+                        if results['success']:
+                            incoming_object = results['object_with_photo_fields']
+
+                    incoming_object.save()
+            # elif hasattr(incoming_object, 'ballotpedia_photo_url_is_broken') \
+            #         and not incoming_object.ballotpedia_photo_url_is_broken:
+            #     incoming_object.ballotpedia_photo_url_is_broken = True
+            #     incoming_object.save()
+
+        # link_is_broken = results.get('http_response_code') == 404
+        is_placeholder_photo = results.get('is_silhouette')
+        if is_placeholder_photo:
+            success = False
+            # status += results['status']
+            status += "IS_PLACEHOLDER_PHOTO "
+            logger.info("Placeholder/Silhouette: " + photo_url)
+            if add_messages:
+                messages.add_message(
+                    request, messages.ERROR,
+                    'Failed to retrieve Ballotpedia picture:  The Ballotpedia URL is for placeholder/Silhouette image.')
+            # Create a record denoting that we have retrieved from Ballotpedia for this candidate
+            if is_candidate:
+                save_results_history = remote_request_history_manager.create_remote_request_history_entry(
+                    kind_of_action=RETRIEVE_POSSIBLE_BALLOTPEDIA_PHOTOS,
+                    google_civic_election_id=google_civic_election_id,
+                    candidate_campaign_we_vote_id=incoming_object.we_vote_id,
+                    number_of_results=1,
+                    status="CANDIDATE_BALLOTPEDIA_URL_IS_PLACEHOLDER_SILHOUETTE:" + str(photo_url))
+        elif results['photo_url_found']:
+            # Success!
+            logger.info("Queried URL: " + ballotpedia_page_url + " ==> " + photo_url)
+            if add_messages:
+                messages.add_message(request, messages.INFO, 'Ballotpedia photo retrieved.')
+            if save_to_database:
+                if is_candidate or is_politician:
+                    results = save_image_to_candidate_table(
+                        candidate=incoming_object,
+                        image_url=photo_url,
+                        source_link=ballotpedia_page_url,
+                        url_is_broken=False,
+                        kind_of_source_website=IMAGE_SOURCE_BALLOTPEDIA)
+                    if results['success']:
+                        ballotpedia_photo_saved = True
+                    # When saving to candidate object, update:
+                    # we_vote_hosted_profile_facebook_image_url_tiny
+                else:
+                    results = save_image_to_organization_table(
+                        incoming_object, photo_url, ballotpedia_page_url, False, IMAGE_SOURCE_BALLOTPEDIA)
+                    if results['success']:
+                        ballotpedia_photo_saved = True
+
+        if not ballotpedia_photo_saved:
+            success = False
+            status += results['status']
+            status += "SAVE_BALLOTPEDIA_IMAGE_TO_CANDIDATE_TABLE_FAILED "
+        else:
+            status += "SAVED_BALLOTPEDIA_IMAGE "
+            # Create a record denoting that we have retrieved from Ballotpedia for this candidate
+
+            if is_candidate:
+                save_results_history = remote_request_history_manager.create_remote_request_history_entry(
+                    kind_of_action=RETRIEVE_POSSIBLE_BALLOTPEDIA_PHOTOS,
+                    google_civic_election_id=google_civic_election_id,
+                    candidate_campaign_we_vote_id=incoming_object.we_vote_id,
+                    number_of_results=1,
+                    status="CANDIDATE_BALLOTPEDIA_URL_PARSED_HTTP:" + ballotpedia_page_url)
+    else:
+        success = False
+        status += results['status']
+        status += "GET_BALLOTPEDIA_FAILED "
+
+        if add_messages:
+            if len(results.get('clean_message')) > 0:
+                messages.add_message(request, messages.ERROR, results.get('clean_message'))
+            else:
+                messages.add_message(
+                    request, messages.ERROR, 'Ballotpedia photo NOT retrieved (2). status: ' + results.get('status'))
+
+    results = {
+        'success': success,
+        'status': status,
+    }
+    return results
 
 def retrieve_ballotpedia_candidates_by_district_from_api(google_civic_election_id, state_code="",
                                                          only_retrieve_if_zero_candidates=False):
