+++ templates/candidate/candidate_list.html
@@ -281,6 +281,9 @@
         </li>
         <li>Prepare: <a href="{% url 'candidate:update_profile_image_background_color' %}?candidate_year={{ show_this_year_of_candidates }}&state_code={{ state_code }}" >
             Update 5,000 Candidates' background colors &#40;may take 10 minutes&#41;</a></li>
+        <li><a href="{% url 'ballotpedia:bulk_retrieve_ballotpedia_photos' %}?google_civic_election_id={{ google_civic_election_id }}&state_code={{ state_code }}&page={{ current_page_number }}"
+               target="_blank" >
+            Retrieve Ballotpedia Photos for up to <strong>{{ballotpedia_urls_without_picture_urls}}</strong> Candidates <span class="glyphicon glyphicon-new-window"></span></a> (If the count does not decrease with each click, you are done for this election)</li>
     </ul>
 {% endif %}
